style(msvc): fix C4244/C4267 in every remaining 3-site file (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226; [NotesLoader]-tagged
tests (587/34) re-checked identical.

19 files, 57 sites: WheelNotifyIcon.cpp, UnlockManager.cpp, Steps.cpp
(SS5-protected, characterization-only Lua binding), ScreenSelectCharacter.cpp,
ScreenSelect.cpp, Screen.cpp, RageThreads.cpp,
RageSoundReader_ChannelSplit.cpp, RageFileDriverReadAhead.cpp,
RageFileDriverDirect.cpp, PlayerStageStats.cpp, OptionRowHandler.cpp,
NotesLoaderSSC.cpp (SS5-protected, characterization-only), LightsManager.cpp,
InputQueue.cpp, GameManager.cpp, ArrowEffects.cpp, ActorFrame.cpp,
ScreenOptionsEditCourse.cpp (a genuinely separate file from
ScreenOptions.cpp/Course.cpp, never touched before).

Every site is a size_t/int64_t/double/lua_Integer-family value
narrowed into a smaller int/unsigned/float/lua_Number type at a
wrap() call, a reverse-loop bound, an iterator-diff, or a Lua
push/rawset binding -- same patterns as the whole sweep.

+++ src/ArrowEffects.cpp
@@ -438,7 +438,7 @@ static float GetDisplayedBeat( const PlayerState* pPlayerState, float beat )
 {
 	// do a binary search here
 	const std::vector<CacheDisplayedBeat> &data = pPlayerState->m_CacheDisplayedBeat;
-	int max = data.size() - 1;
+	int max = static_cast<int>(data.size()) - 1;
 	int l = 0, r = max;
 	while( l <= r )
 	{
@@ -833,7 +833,7 @@ float ArrowEffects::GetXPos( const PlayerState* pPlayerState, int iColNum, float
 				{
 					// find the middle, and split based on iColNum
 					// it's unknown if this will work for routine.
-					const int iMiddleColumn = std::floor(pStyle->m_iColsPerPlayer/2.0f);
+					const int iMiddleColumn = static_cast<int>(std::floor(pStyle->m_iColsPerPlayer/2.0f));
 					if( iColNum > iMiddleColumn-1 )
 						fPixelOffsetFromCenter += fEffects[PlayerOptions::EFFECT_XMODE]*-(fYOffset);
 					else
